requirements.txt
feedparser>=6.0.0
newspaper3k>=0.2.8
pyyaml>=6.0
edge-tts>=6.1.0
lxml>=4.9.0
tqdm>=4.64.0
ollama>=0.1.0
requests>=2.28.0
openai>=1.0.0
scikit-learn>=1.0.0
aiohttp>=3.8.0
textblob>=0.15.3
spacy>=3.0.0
gradio>=3.0.0
scikit-learn 
sentence-transformers 
umap-learn  
transformers 
accelerate
geopy
fastapi>=0.68.0
uvicorn>=0.15.0
